api/requirements.txt
fastapi
azure-cosmos
azure-identity
uvicorn
python-dotenv
azure-monitor-opentelemetry
aiohttp